# backend/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "藏药小程序API",
    "description": "藏药电商小程序后端API接口文档",
    "version": "1.0.0",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "本地开发环境"
    }
  ],
...
  "paths": {
    "/user/login": {
      "post": {
        "tags": ["用户模块"],
        "summary": "用户登录",
        "description": "微信小程序登录（Mock版本）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "code": {
                    "type": "string",
                    "description": "微信登录code"
                  }
                },
                "required": ["code"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/info": {
      "get": {
        "tags": ["用户模块"],
        "summary": "获取用户信息",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserInfoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "post": {
        "tags": ["用户模块"],
        "summary": "更新用户信息",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {"type": "string"},
                  "phone": {"type": "string"},
                  "avatar": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "LoginResponse": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "message": {"type": "string"},
          "data": {
            "type": "object",
            "properties": {
              "token": {"type": "string"},
              "userInfo": {"type": "object"}
            }
          }
        }
      },
      "UserInfoResponse": {
        "type": "object",
        "properties": {
          "code": {"type": "integer"},
          "data": {
            "type": "object",
            "properties": {
              "id": {"type": "integer"},
              "nickName": {"type": "string"},
              "avatar": {"type": "string"},
              "phone": {"type": "string"}
            }
          }
        }
      },
...
    }
  }
}